+++ run.py
@@ -26,8 +26,14 @@ def create_app(config):
 
     # register view blueprints
     from ras_party.views.party_view import party_view
+    from ras_party.views.business_view import business_view
+    from ras_party.views.respondent_view import respondent_view
+    from ras_party.views.account_view import account_view
     from ras_party.views.info_view import info_view
     app.register_blueprint(party_view, url_prefix='/party-api/v1')
+    app.register_blueprint(account_view, url_prefix='/party-api/v1')
+    app.register_blueprint(business_view, url_prefix='/party-api/v1')
+    app.register_blueprint(respondent_view, url_prefix='/party-api/v1')
     app.register_blueprint(info_view)
 
     CORS(app)
